+++ packages/pangea-node-sdk/tests/integration/file_scan.test.ts
@@ -168,7 +168,7 @@ it("File Scan get url and put upload", async () => {
   };
   let response = await fileScan.requestUploadURL(request);
 
-  const url = response.accepted_result?.put_url || "";
+  const url = new URL(response.accepted_result?.put_url || "");
 
   const uploader = new FileScanUploader();
   await uploader.uploadFile(
@@ -221,7 +221,7 @@ it("File Scan get url and post upload", async () => {
     return;
   }
 
-  const url = response.accepted_result?.post_url || "";
+  const url = new URL(response.accepted_result?.post_url || "");
   const file_details = response.accepted_result?.post_form_data;
 
   const uploader = new FileScanUploader();