+++ packages/pangea-node-sdk/CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Support for Node.js v18. The minimum supported version is now Node.js v20.
 - Deprecated APIs like `PangeaConfig.configID`, `Vault.AsymmetricAlgorithm.RSA`,
   and `Vault.SymmetricAlgorithm.AES`.
 - AI Guard: `llm_info` and `llm_input`.